Revert "Scrum 88 music and sound effects"
