+++ .gitignore
@@ -39,25 +39,3 @@ sysinfo.txt
 Icon?
 ehthumbs.db
 Thumbs.db
-Gamedev/Logs/AssetImportWorker2.log
-Gamedev/Logs/AssetImportWorker3-prev.log
-Gamedev/Logs/AssetImportWorker3.log
-Gamedev/Logs/AssetImportWorker4-prev.log
-Gamedev/Logs/AssetImportWorker5-prev.log
-Gamedev/Logs/AssetImportWorker6-prev.log
-Gamedev/Logs/AssetImportWorker7-prev.log
-Gamedev/Logs/shadercompiler-AssetImportWorker0.log
-Gamedev/Logs/AssetImportWorker1.log
-Gamedev/Logs/AssetImportWorker1-prev.log
-Gamedev/Logs/AssetImportWorker0.log
-Gamedev/Logs/AssetImportWorker0-prev.log
-Gamedev/Logs/AssetImportWorker2-prev.log
-Gamedev/Logs/AssetImportWorker1.log
-Gamedev/Logs/AssetImportWorker1-prev.log
-Gamedev/Logs/AssetImportWorker0.log
-Gamedev/Logs/AssetImportWorker0-prev.log
-Gamedev/Logs/AssetImportWorker0-prev.log
-Gamedev/Logs/AssetImportWorker0.log
-Gamedev/Logs/AssetImportWorker1-prev.log
-Gamedev/Logs/AssetImportWorker1.log
-Gamedev/Logs/AssetImportWorker0-prev.log
