--- Gamedev/Assets/Scenes/GameScene/NavMesh-Grass 2.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0636ed9e7f4fd6e4e905ce6c03168164
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
