+++ Gamedev/Assets/Scenes/GameScene_Profiles/dayProfile.asset
@@ -68,31 +68,31 @@ MonoBehaviour:
     value: 1
   temperature:
     overrideState: 1
-    value: 0
+    value: 6
   tint:
-    overrideState: 0
+    overrideState: 1
     value: 0
   colorFilter:
-    overrideState: 0
+    overrideState: 1
     value: {r: 1, g: 0.90606517, b: 0.72264147, a: 1}
   hueShift:
     overrideState: 0
     value: 0
   saturation:
     overrideState: 1
-    value: 39.43
+    value: 34
   brightness:
     overrideState: 0
     value: 14
   postExposure:
     overrideState: 1
-    value: 2.02
+    value: 2.5
   contrast:
     overrideState: 1
-    value: 29.63
+    value: -24
   mixerRedOutRedIn:
-    overrideState: 1
-    value: 33.9
+    overrideState: 0
+    value: 100
   mixerRedOutGreenIn:
     overrideState: 0
     value: 0
@@ -104,22 +104,22 @@ MonoBehaviour:
     value: 0
   mixerGreenOutGreenIn:
     overrideState: 0
-    value: 120
+    value: 100
   mixerGreenOutBlueIn:
     overrideState: 0
     value: 0
   mixerBlueOutRedIn:
     overrideState: 0
-    value: 13
+    value: 0
   mixerBlueOutGreenIn:
-    overrideState: 1
-    value: 27
+    overrideState: 0
+    value: 0
   mixerBlueOutBlueIn:
-    overrideState: 1
-    value: 132
+    overrideState: 0
+    value: 100
   lift:
-    overrideState: 1
-    value: {x: 1, y: 1, z: 1, w: -0.11794439}
+    overrideState: 0
+    value: {x: 1, y: 1, z: 1, w: 0}
   gamma:
     overrideState: 0
     value: {x: 1, y: 1, z: 1, w: 0}
@@ -1338,13 +1338,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 48a79b01ea5641d4aa6daa2e23605641, type: 3}
   m_Name: Bloom
   m_EditorClassIdentifier: 
-  active: 1
+  active: 0
   enabled:
     overrideState: 1
     value: 1
   intensity:
     overrideState: 1
-    value: 0.05
+    value: 1
   threshold:
     overrideState: 1
     value: 0
