+++ Gamedev/Assets/Scenes/GameScene_Profiles/nightProfile.asset
@@ -18,7 +18,7 @@ MonoBehaviour:
     value: 1
   intensity:
     overrideState: 1
-    value: 0.75
+    value: 1
   threshold:
     overrideState: 0
     value: 1
@@ -30,7 +30,7 @@ MonoBehaviour:
     value: 65472
   diffusion:
     overrideState: 1
-    value: 10
+    value: 4
   anamorphicRatio:
     overrideState: 0
     value: 0
@@ -100,64 +100,64 @@ MonoBehaviour:
     value: 1
   temperature:
     overrideState: 1
-    value: 0
+    value: -25
   tint:
     overrideState: 0
-    value: 10
+    value: 0
   colorFilter:
-    overrideState: 1
-    value: {r: 0.037155293, g: 0, b: 1.3584905, a: 1}
+    overrideState: 0
+    value: {r: 1, g: 1, b: 1, a: 1}
   hueShift:
     overrideState: 0
     value: 0
   saturation:
     overrideState: 1
-    value: 3.8
+    value: 40
   brightness:
     overrideState: 1
     value: -50
   postExposure:
     overrideState: 1
-    value: 5.26
+    value: -2
   contrast:
-    overrideState: 1
-    value: 31.5
+    overrideState: 0
+    value: 0
   mixerRedOutRedIn:
-    overrideState: 1
-    value: -52
+    overrideState: 0
+    value: 100
   mixerRedOutGreenIn:
-    overrideState: 1
-    value: 5
+    overrideState: 0
+    value: 0
   mixerRedOutBlueIn:
     overrideState: 0
-    value: 2
+    value: 0
   mixerGreenOutRedIn:
     overrideState: 0
-    value: 4
+    value: 0
   mixerGreenOutGreenIn:
-    overrideState: 1
-    value: 25
+    overrideState: 0
+    value: 100
   mixerGreenOutBlueIn:
     overrideState: 0
-    value: 38
+    value: 0
   mixerBlueOutRedIn:
     overrideState: 0
-    value: 73
+    value: 0
   mixerBlueOutGreenIn:
-    overrideState: 1
-    value: 43
+    overrideState: 0
+    value: 0
   mixerBlueOutBlueIn:
-    overrideState: 1
-    value: 111
+    overrideState: 0
+    value: 100
   lift:
-    overrideState: 1
-    value: {x: 1, y: 0.9882925, z: 0.9847178, w: -0.3220657}
+    overrideState: 0
+    value: {x: 1, y: 1, z: 1, w: 0}
   gamma:
     overrideState: 0
-    value: {x: 1, y: 1, z: 1, w: -0.1610499}
+    value: {x: 1, y: 1, z: 1, w: 0}
   gain:
     overrideState: 0
-    value: {x: 1, y: 1, z: 1, w: 1}
+    value: {x: 1, y: 1, z: 1, w: 0}
   masterCurve:
     overrideState: 0
     value:
@@ -1378,14 +1378,14 @@ MonoBehaviour:
     overrideState: 0
     value: 0
   color:
-    overrideState: 1
-    value: {r: 0.20817567, g: 0.039248843, b: 0.3962264, a: 1}
+    overrideState: 0
+    value: {r: 0, g: 0, b: 0, a: 1}
   center:
     overrideState: 0
     value: {x: 0.5, y: 0.5}
   intensity:
     overrideState: 1
-    value: 0.91
+    value: 0.4
   smoothness:
     overrideState: 0
     value: 0.2
@@ -1400,6 +1400,43 @@ MonoBehaviour:
     value: {fileID: 0}
     defaultState: 1
   opacity:
+    overrideState: 0
+    value: 1
+--- !u!114 &-3580289678369907284
+MonoBehaviour:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b3f6f3f7c722b4544b97e3c75840aa33, type: 3}
+  m_Name: AutoExposure
+  m_EditorClassIdentifier: 
+  active: 0
+  enabled:
+    overrideState: 1
+    value: 1
+  filtering:
+    overrideState: 0
+    value: {x: 53.999985, y: 99}
+  minLuminance:
+    overrideState: 0
+    value: 0
+  maxLuminance:
+    overrideState: 0
+    value: 0
+  keyValue:
+    overrideState: 1
+    value: 1
+  eyeAdaptation:
+    overrideState: 0
+    value: 0
+  speedUp:
+    overrideState: 0
+    value: 2
+  speedDown:
     overrideState: 0
     value: 1
 --- !u!114 &11400000
@@ -1417,4 +1454,5 @@ MonoBehaviour:
   settings:
   - {fileID: -7358687046336904684}
   - {fileID: -5044330839455133485}
+  - {fileID: -3580289678369907284}
   - {fileID: -8380701970544134533}
