--- Gamedev/Assets/Sound Effects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a3cef06830a87c4469ef78a7e09a6c00
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
