--- Gamedev/Assets/Sound Effects/background-royalty-free-music-emotional-piano-305670.mp3.meta
@@ -1,23 +1,0 @@
-fileFormatVersion: 2
-guid: 7bd9ffc10b8a5f24e90ae8b026c75ba3
-AudioImporter:
-  externalObjects: {}
-  serializedVersion: 7
-  defaultSettings:
-    serializedVersion: 2
-    loadType: 0
-    sampleRateSetting: 0
-    sampleRateOverride: 44100
-    compressionFormat: 1
-    quality: 1
-    conversionMode: 0
-    preloadAudioData: 0
-  platformSettingOverrides: {}
-  forceToMono: 0
-  normalize: 1
-  loadInBackground: 0
-  ambisonic: 0
-  3D: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
