--- Gamedev/Assets/SoundEffects.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5ecb3ccb54437584daa2c19b7282ffb2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
