+++ Gamedev/Assets/UI/UIController.cs.meta
@@ -1,0 +1,15 @@
+fileFormatVersion: 2
+guid: 2806f85e9f3f1a24db90e4dc72097b59
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences:
+  - shopPanel: {fileID: 7128684620851026400, guid: 38629499230ef8048a85f0492fc62fd1, type: 3}
+  - inventoryPanel: {fileID: 3870177152600446569, guid: 90066252a96a623489d4ee0153488968, type: 3}
+  - coinUI: {fileID: 1255623348378946956, guid: 99d56c1e561884f40ae708cfda35d760, type: 3}
+  - placementSystem: {instanceID: 0}
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
