+++ Gamedev/Assets/scrip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54b42fb76023dfb47a7e7ef1aad707fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
