--- Gamedev/Assets/scripts/UIController.cs.meta
@@ -1,17 +1,0 @@
-fileFormatVersion: 2
-guid: 2806f85e9f3f1a24db90e4dc72097b59
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - shopPanel: {instanceID: 0}
-  - inventoryPanel: {instanceID: 0}
-  - coinsLabelShop: {instanceID: 0}
-  - audioSFX: {instanceID: 0}
-  - BG_NightSFX: {instanceID: 0}
-  - BG_DaySFX: {instanceID: 0}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
